Big change
Code refactor, new folders added and files moved. Updated the overlay to utilise an enum for tracking state changes in the star map and within the overlay (Reduces the number of required props and elevates any click handlers to the top level). Form added to the overlay for adding new objects, not yet hooked up to the server.

Yet to be done: Refactor of both star and celestial body data structures to match. We will treat them as graph nodes to ensure that planets, stars, asteroids, etc can exist as standalone items (as root nodes) if required