Updated star endpoints
Added function to add a full list of stars to the database. This only needed to be run once to transfer local star data from a JSON file, to MongoDB. The function will stay for now but currently has no use.

Function to retrieve all stars was updated and is now functional.