Installed typescript map
As JSON cannot currently address maps in an appropriate way (keys cannot be objects), I added typescript-map. Depending on how data needs to be stored and represented on the front-end (i.e. the star map as a graph) it will be handy to have available.

If used, direct conversion to JSON will appear a little ~messy~ as we end up with a list of keys and a list of values. It's (potential) usage comes down to which functions are needed on the front-end, how we want data represented in the database, and what functions can be delegated to the server.