Update to data formats
Big changes:

Star.ts - Defines a class for our star data as a one-to-one of the previous JSONStar interface. As the project settles into MongoDB a bit more fully the attributes here will be renamed to be both clearer, and shared across other object types (Planet, Asteroid, Belt, Ship, etc..)

Stars.tsx - Current placeholder of the desired attribute names which I have left in for now as I decide what goes and what stays. Ultimately earmarked for removal in place of Star.ts

threeJs\Objects\Star.tsx - Currently a threeJS object used in the solar system display. Hopefully, we can merge this in with our Star.ts class and use it as the springboard for how other objects need to be represented.

CelestialBody.tsx - If we can successfully come up with shared attributes for Planet, Star, Asteroid, Belt, etc... CelestialBody.tsx will overrule Star.tsx (and any other similar classes we may define). The reasoning for this is that each ThreeJS object we want to display will ultimately be a defined point in space, that may or may not move, has a name, has a description, and renders. Depending upon a provided enum value (Star | Planet | ... ) the display type may change. For now it remains as a placeholder.

interfaces.ts - Most noticeable change, the addition of the CBType enum which is not yet in use. If the CelestialBody.tsx approach is taken, it's a necessary evil to store the type in transit to the DB anyway.

UniverseGraph.ts - A food for thought representation of the data we currently have. A graph is the most appropriate representation of the Star chart, such that each star will have adjacent or connected stars. Each Star will then exist as a root node for it's own solar system, where children could be of CBType (refer to above). This poses some interesting cases with that of n-ary star systems, where the root node is just "the system" and not necessarily a star. It just so happens that all nodes we currently have are stars.

The more we manipulate and massage the data and start creating data the clearer all of this will become