[GREY-15] Adapted Node constructor
Adapted the Node constructor and attributes to be in-line with the MongoDB representation. An interface was added to aid in type checking when generating and submitting forms.

I.e. the form will be used to add all constructor parameters (INode), creating a new Node and finally sending the new Node to the server for storage.

This is not yet final.