[GREY-1] Added loader to pointsOfInterest
Also updated the loader to have highlight values, currently it only handles black and white which is what we are currently sticking to for page backgrounds. In future this will need to update along with the colour scheme. Regardless, values should be used into constant values at the root of the application.

NOTE: If refreshing the page, background colour defaults to white. Will need a check to see what the current page is in App.tsx to handle this.