[GREY-19] Updated search layout and focus
Moved search bar into the centre of the screen for the time being to keep it separate from the current star information to the right side.

Added onBlur and onFocus listeners which we should have had before.

Will mark as done, any additional changes to UI or functionality extend beyond the implementation of this ticket.