[GREY-33] Added async/await to stars route
Server was returning a response prematurely before the promise had been resolved. Async/await added to handle this. We will not yet add in authentication but I look forward to implementing it to ensure we can store different stories in the database and different levels of access for each project (i.e. edit or read-only).