[GREY-33, GREY-6] Cleaned star routes
References to the old dbo constant were removed to opt for the async/await try/catch structure instead. The code compiles however not all functions have been tested yet. Not all functions are yet in use on the frontend so will be updated as needed. Error codes and messages will be improved if needed as well.