[GREY-32] Classes moved to models folder
The server is going to import models from the frontend to avoid needing to recreate the data structures in multiple locations. This will allow us to validate data received from the client, type check data before sending it to the client, and structure the MongoDB schemas around the same data structures as well whilst removing multiple points of failure.

It will look inappropriate importing files from the frontend/client but limits the class and interface definitions to one location.