[GREY-32] Added base star schema
I have started adding the star schema, however I am uncertain of the best way to structure it or if the schema is even necessary when the server has access to the types. Given that the server handles storage and retrieval the schema will remain unchanged if we validate types outside of the database.