[GREY-12] Created celestial body class
The new CelestialBody class was created to be used as the main model for the server. The ./src/components/threejs/Objects/CelestialBody.tsx will need to be migrated, moved or deleted.

This class will also be used in place of the CBState and CBProps defined in interfaces.ts. Following this commit will be a bit of a cleanup around leftover interfaces, classes and types.