[GREY-6] CORS options added
Currently addNode is not resolving and being blocked by CORS. The added options do not fix this yet but will be updated as we go. Status codes also now need to be handled on the front end, prompting the addition of a possible helper function to address typical status code resolutions.