[GREY-36] Fixed star select performance issue
Updating the state of a screen position every frame was too heavy. The code is a little less clean but more functional. Bringing in React-redux may be the cleaner solution as there could be other elements and positions to update later down the track. Other alternatives will be considered after more urgent issues are handled.