[GREY-12] Cleanup of SolarSystem.tsx
Gravity warp and fabric was moved into it's own file. Solar system now takes in a node so that we have access to the related children for whichever node was selected. Currently children should be passed in with the node object and should not require any additional requests to the server to be made.